backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# Triton client
tritonclient[grpc]==2.44.0
grpcio==1.63.0

# ML
torch==2.3.0+cu121
torchvision==0.18.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

transformers==4.41.2
accelerate==0.30.1
sentence-transformers==3.0.1
datasets==2.19.1

# Vector store
faiss-gpu-cu12==1.7.4

# ONNX
onnxruntime-gpu==1.18.0
onnx==1.16.0

# PDF parsing
PyMuPDF==1.24.3

# Utils
numpy==1.26.4
huggingface-hub==0.23.0
